remove relay from sdk
