+++ packages/relay/package.json
@@ -52,6 +52,7 @@
     "@chainsafe/libp2p-gossipsub": "^13.1.0",
     "@noble/hashes": "^1.3.2",
     "@waku/core": "0.0.31",
+    "@waku/sdk": "0.0.27",
     "@waku/interfaces": "0.0.26",
     "@waku/proto": "0.0.8",
     "@waku/utils": "0.0.19",
