+++ packages/sdk/src/create/index.ts
@@ -1,2 +1,2 @@
 export { createLightNode } from "./create.js";
-export { defaultLibp2p } from "./libp2p.js";
+export { defaultLibp2p, createLibp2pAndUpdateOptions } from "./libp2p.js";
