+++ packages/sdk/src/index.ts
@@ -9,7 +9,11 @@ export { utf8ToBytes, bytesToUtf8 } from "@waku/utils/bytes";
 
 export * from "./waku.js";
 
-export { createLightNode, defaultLibp2p } from "./create/index.js";
+export {
+  createLightNode,
+  defaultLibp2p,
+  createLibp2pAndUpdateOptions
+} from "./create/index.js";
 export { wakuLightPush } from "./protocols/light_push.js";
 export { wakuFilter } from "./protocols/filter.js";
 export { wakuStore } from "./protocols/store.js";
@@ -17,4 +21,3 @@ export { wakuStore } from "./protocols/store.js";
 export * as waku from "@waku/core";
 export * as utils from "@waku/utils";
 export * from "@waku/interfaces";
-export * as relay from "@waku/relay";
