+++ packages/tests/package.json
@@ -77,6 +77,7 @@
     "@waku/discovery": "*",
     "@waku/message-encryption": "*",
     "@waku/sdk": "*",
+    "@waku/relay": "*",
     "allure-commandline": "^2.27.0",
     "allure-mocha": "^2.9.2",
     "chai": "^4.3.10",
