+++ packages/tests/tests/wait_for_remote_peer.node.spec.ts
@@ -1,8 +1,8 @@
 import { waitForRemotePeer } from "@waku/core";
 import type { LightNode, RelayNode } from "@waku/interfaces";
 import { Protocols } from "@waku/interfaces";
+import { createRelayNode } from "@waku/relay";
 import { createLightNode } from "@waku/sdk";
-import { createRelayNode } from "@waku/sdk/relay";
 import { expect } from "chai";
 
 import {
